+++ core/scene/Camera.cpp
@@ -41,7 +41,7 @@ namespace LSIS {
 
 	glm::mat4 Camera::GetViewProjectionMatrix() const
 	{
-		float aspect = m_resolution.x / m_resolution.y;
+		float aspect = static_cast<float>(m_resolution.x) / static_cast<float>(m_resolution.y);
 		glm::mat4 P = glm::perspective(glm::radians(m_FOV), aspect, 0.1f, 100.0f);
 		glm::mat4 V = m_transform.GetModelMatrix();
 		return P * glm::inverse(V);
